refactor: ChessController (move, update 분리)
